requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
aiohttp-debugtoolbar==0.5.0
aiohttp-jinja2==1.2.0
aioredis==1.3.1
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
hiredis==1.0.1
idna==2.8
Jinja2==2.10.3
libsass==0.19.4
MarkupSafe==1.1.1
multidict==4.5.2
pbr==5.4.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.10
sqlalchemy-migrate==0.13.0
sqlparse==0.3.0
Tempita==0.5.2
urllib3==1.25.7
yarl==1.3.0
